Squashed commit of the following:
commit f9b4fff5604fc8796c55fe7ba74b0e48b84e3f00
Author: EXT Leiva, Mauro <Mauro.Leiva@molinosagro.com.ar>
Date:   Mon Jan 20 15:53:10 2025 +0000

    Merged PR 2658: Merged PR 2655: ARMOA005-2521 Habilitar AFIP en Producción

    Merged PR 2655: ARMOA005-2521 Habilitar AFIP en Producción

    ARMOA005-2521 Habilitar AFIP en Producción

commit fef5edc0263e0caa06afb52ac93ddbf124af2a15
Author: EXT Leiva, Mauro <Mauro.Leiva@molinosagro.com.ar>
Date:   Mon Jan 20 15:52:16 2025 +0000

    Merged PR 2657: ARMOA005-2510 Fix orden calculos lineas de embarque

    ARMOA005-2510 Fix orden calculos lineas de embarque

commit 29f07d96d4f4225905d55a77d7f666c6d2158948
Author: EXT Leiva, Mauro <Mauro.Leiva@molinosagro.com.ar>
Date:   Mon Jan 20 15:51:48 2025 +0000

    Merged PR 2656: ARMOA005-2503 Lineas de embarque duplicadas

    ARMOA005-2503 Lineas de embarque duplicadas

    - Se corrigió un error que duplicaba los tipos de linea y materiales al realizar un guardado general con nuevas lineas

commit ab389397990025afc896f7971d0db19786fb0c2a
Merge: 8b7402c3 c24f8b6e
Author: Oscar De Pierola <odepierola@baufest.com>
Date:   Fri Jan 17 09:20:28 2025 -0500

    Merge remote-tracking branch 'dev.azure.com/hotfix/mpuca/ARMOA005-2516' into release/1.19.1.0

commit 8b7402c3d6f015578c22c57681f5827c744109d6
Merge: 640ed4d0 b2bb3845
Author: Oscar De Pierola <odepierola@baufest.com>
Date:   Fri Jan 17 09:19:32 2025 -0500

    Merge remote-tracking branch 'dev.azure.com/hotfix/mpuca/ARMOA005-2507' into release/1.19.1.0

commit 640ed4d0027c59dd28cec501e795ca97286f06d5
Author: Mauro Leiva <mleiva@baufest.com>
Date:   Thu Jan 16 10:03:00 2025 -0300

    ARMOA005-2505 Error en calculo de ritmos

    - Se cambiaron los labels que decían "por hora por mano" a solo "por hora"

commit c24f8b6e33b800477ea39291cf3aeeded9d8de19
Author: BAUNET\mpuca <mpuca@baufest.com>
Date:   Thu Jan 16 09:28:10 2025 -0300

    Fix agrandar longitud campo stowed para recibo.

commit b2bb384546df6929d6ccc8284f7a59db7d615ee9
Author: BAUNET\mpuca <mpuca@baufest.com>
Date:   Wed Jan 15 10:44:14 2025 -0300

    Fix permitir decimales en planilla embarque liquidos, y fix validacion horarios exportador.

commit a079fb09eb7aae6e1423d22fea950de69f13bffe
Author: Mauro Leiva <mleiva@baufest.com>
Date:   Mon Jan 13 17:14:03 2025 -0300

    ARMOA005-2505 Error en calculo de ritmo neto

    - Se corrigió unbug donde se restaba el tiempo de las cargas normales para el cálculo del neto
    - Se cambió el label del reloj de "Ritmo de carga" para quitarle el texto "por mano"

+++ Molinos.Scato.Database/dbo/Tables/ModuloDeCargaPlanillaDeEmbarque.sql
@@ -7,7 +7,7 @@
 	[TanqueDeAbordo]			nvarchar(50),
 	[Destino_Id]				INT NULL,
 	[Tk]						nvarchar(10),
-	[Tn]						int,
+	[Tn]						decimal(18,3),
 	[Cantidad]					int,
 	[MaterialPuerto_Id]			INT NOT NULL,
 	[FechaComienzoCarga]		datetime,
